Created package quest in main and test folder to tidy up project structure. Moved quest classes and tests in to package quest
