+++ Roguelike/src/main/java/se/su/dsv/inte/quest/QuestObjective.java
@@ -1,4 +1,4 @@
-package se.su.dsv.inte;
+package se.su.dsv.inte.quest;
 
 public abstract class QuestObjective {
     public abstract String getStatusMessage();
