+++ Roguelike/src/test/java/se/su/dsv/inte/quest/MonsterSlayingObjectiveTest.java
@@ -1,6 +1,8 @@
-package se.su.dsv.inte;
+package se.su.dsv.inte.quest;
 
 import org.junit.Test;
+import se.su.dsv.inte.quest.MonsterSlayingObjective;
+import se.su.dsv.inte.quest.QuestObjective;
 
 import static org.junit.Assert.*;
