+++ Roguelike/src/test/java/se/su/dsv/inte/quest/QuestObjectiveTest.java
@@ -1,4 +1,4 @@
-package se.su.dsv.inte;
+package se.su.dsv.inte.quest;
 
 import org.junit.Test;
